requirements.txt
torch
torchvision
numpy
pandas
opencv-python
scikit-learn
lightgbm
tqdm
Pillow
